api/requirements.txt
fastapi
pydantic
python-dotenv
uvicorn
langchain
openai
notion

